Create functions that receive repos and place them
I created a file called github-repo-project-cards.js. This file contains
three functions; a function that makes an API call to get all repos from
a particular github user, a function that takes that data and creates
project cards to be appended to the dom, and a controller function that
controls the flow of the first two functions.

I realized, however, that the injected cards are not being styled like
the hard coded cards are. After doing some digging, it seems as though
the scripts found at the end of the index.html file are jquery plugins
that control the styling of the cards. I want to separate the logic of
these jquery scripts, create a new controller function found in a
separate file, and have that file control the flow of all of these
functions. I believe that the jquery scripts in charge of styling those
cards are not being called after the new cards are injected into that
space.