Migrate JQuery styling scripts to new feature file
I found out the reason for the previous issues with the newly injected
project cards. The projects were being injected into the portfolio
section after the JQuery scripts were executed. These JQuery scripts,
originally located at the bottom of the index.html file, are responsible
for the styling of the portfolio section. To solve the issue, I migrated
this JQuery to the the feature js file, wrapped it all in a new
function, and execute said function only after all projects are injected
into the portfolio section.

There are still a few issues however;

1. When refreshing the page, sometimes, the project cards extend into
the next section.

2. For some reason, the C++ and C# filter buttons don't yield any
projects.